% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\name{anova.refitME}
\alias{anova.refitME}
\title{anova.refitME}
\usage{
\method{anova}{refitME}(object, ...)
}
\arguments{
\item{object}{: fitted model objects of class refitME.}

\item{...}{: further arguments passed through to \code{glm} or \code{gam}.}
}
\value{
\code{ranova.refitME} produces output identical to \code{anova.gam}.
}
\description{
An ANOVA function for fitted refitME objects.
}
\seealso{
\code{\link{anova.gam}}
}
\author{
Jakub Stoklosa and David I. Warton.
}
