% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.r
\name{anova_MCEMfit_glm}
\alias{anova_MCEMfit_glm}
\title{An ANOVA function for fitted \code{MCEMfit_glm} objects}
\usage{
anova_MCEMfit_glm(object, ..., dispersion = NULL, test = NULL)
}
\arguments{
\item{object}{: fitted model objects of class \code{MCEMfit_glm}.}

\item{...}{: further arguments passed through to \code{glm}.}

\item{dispersion}{: the dispersion parameter for the fitting family. By default it is obtained from the object(s).}

\item{test}{: a character string, (partially) matching one of "\code{Chisq}", "\code{LRT}", "\code{Rao}", "\code{F}" or "\code{Cp}".}
}
\value{
\code{anova_MCEMfit_glm} produces output identical to \code{anova.glm}.
}
\description{
An ANOVA function for fitted \code{MCEMfit_glm} objects.
}
\seealso{
\code{\link{anova.glm}}
}
\author{
Jakub Stoklosa, Wen-Han Hwang and David I. Warton.
}
