% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\docType{data}
\name{Framinghamdata}
\alias{Framinghamdata}
\title{The Framingham heart study data set.}
\format{A data set contains: 5 columns with 1615 observations. The columns are defined as follows:
\describe{
 \item{\code{Y}}{The response binary variable of whether the patient had a heart attack.}
 \item{\code{z1}}{Cholesterol of patient.}
 \item{\code{z2}}{Age of patient.}
 \item{\code{z3}}{Whether the patient smokes.}
 \item{\code{w1}}{The error contaminated variable - these are mean scores.}
}}
\source{
See Carroll \emph{et al.} (2006) for full details of the data and study. Also, see \url{https://github.com/JakubStats/refitME} for an RMarkdown tutorial of an example that uses the data.
}
\usage{
Framinghamdata
}
\description{
The Framingham heart study data set.
}
\examples{
## Load the data.

data(Framinghamdata)
}
\references{
Carroll, R. J., Ruppert, D., Stefanski, L. A., and Crainiceanu, C. M. (2006). \emph{Measurement Error in Nonlinear Models: A Modern Perspective.} 2nd Ed. London: Chapman \& Hall/CRC.
}
\keyword{datasets}
