% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\name{logLik.MCEMfit_lm}
\alias{logLik.MCEMfit_lm}
\title{Extract log-Likelihoods for \code{logLik.MCEMfit_lm} model objects}
\usage{
\method{logLik}{MCEMfit_lm}(object, REML = FALSE, ...)
}
\arguments{
\item{object}{: fitted model objects of class refitME.}

\item{REML}{: an optional logical value. If \code{TRUE} the restricted log-likelihood is returned, else, if \code{FALSE}, the log-likelihood is returned. Defaults to \code{FALSE}.}

\item{...}{: further arguments passed through to \code{lm}.}
}
\value{
\code{logLik.MCEMfit_lm} produces output identical to \code{logLik}.
}
\description{
Extract log-Likelihoods for \code{refitME} model objects.
}
\seealso{
\code{\link{logLik}}
}
\author{
Jakub Stoklosa and David I. Warton.
}
