% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\name{logLik.refitME}
\alias{logLik.refitME}
\title{Extract log-Likelihoods for \code{refitME} model objects}
\usage{
\method{logLik}{refitME}(object, ...)
}
\arguments{
\item{object}{: fitted model objects of class \code{refitME}.}

\item{...}{: further arguments passed through to \code{lm} or \code{glm}.}
}
\value{
\code{logLik.refitME} produces identical output to \code{logLik} but for \code{refitME} model objects.
}
\description{
Extract log-Likelihoods for \code{refitME} model objects. This function subtracts the entropy term from the observed likelihood.
}
\seealso{
\code{\link{logLik}}
}
\author{
Jakub Stoklosa, Wen-Han Hwang and David I. Warton.
}
