% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\name{logLik_MCEMfit_lm}
\alias{logLik_MCEMfit_lm}
\alias{logLik.refitME}
\title{Extract log-Likelihoods for \code{refitME} model objects}
\usage{
\method{logLik}{refitME}(object, ...)

logLik_MCEMfit_lm(object, REML = FALSE, ...)
}
\arguments{
\item{object}{: fitted model objects of class \code{MCEMfit_lm}.}

\item{...}{: further arguments passed through to \code{lm}.}

\item{REML}{: an optional logical value. If \code{TRUE} the restricted log-likelihood is returned, else, if \code{FALSE}, the log-likelihood is returned. Defaults to \code{FALSE}.}
}
\value{
\code{logLik.refitME} produces output identical to \code{logLik}.

\code{logLik_MCEMfit_lm} produces output identical to \code{logLik}.
}
\description{
Extract log-Likelihoods for \code{efitME} model objects. This function subtracts the entropy term from the observed likelihood.

Extract log-Likelihoods for \code{MCEMfit_lm} model objects. This function subtracts the entropy term from the observed likelihood.
}
\seealso{
\code{\link{logLik}}

\code{\link{logLik}}
}
\author{
Jakub Stoklosa, Wen-Han Hwang and David I. Warton.

Jakub Stoklosa, Wen-Han Hwang and David I. Warton.
}
