% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\name{MCEMfit_CR}
\alias{MCEMfit_CR}
\title{Function for fitting \code{VGAM} capture-recapture (CR) model using the MCEM algorithm}
\source{
See \url{https://github.com/JakubStats/refitME} for an RMarkdown tutorial with examples.
}
\usage{
MCEMfit_CR(
  mod,
  sigma.sq.u,
  sigma.sq.e = 1,
  B = 100,
  epsilon = 1e-05,
  silent = FALSE
)
}
\arguments{
\item{mod}{: a vglm/vgam object (this is the naive fitted CR model). Make sure the first input predictor variables are the selected error-contaminated variable (i.e., the \code{W}'s).}

\item{sigma.sq.u}{: measurement error variance. A scalar if there is only one error-contaminated variable, otherwise this must be stored as a covariance matrix.}

\item{sigma.sq.e}{: variance of the true covariate (X).}

\item{B}{: the number of Monte Carlo replication values (default is set to 50).}

\item{epsilon}{: a set convergence threshold (default is set to 0.00001).}

\item{silent}{: if \code{TRUE}, convergence message is suppressed (default is set to \code{FALSE}).}
}
\value{
\code{MCEMfit_CR} returns model coefficient and population size estimates with standard errors and the effective sample size.
}
\description{
Function for fitting \code{VGAM} capture-recapture (CR) model using the MCEM algorithm where covariates have measurement error.
}
\section{Warning}{

This function is still under development. Currently the function can only fit the CR model used in the manuscript. IT DOES NOT SUPPORT ALL \code{VGAM} families.
}

\references{
Stoklosa, J., Hwang, W-H., and Warton, D.I. \pkg{refitME}: Measurement Error Modelling using Monte Carlo Expectation Maximization in \proglang{R}.
}
\seealso{
\code{\link{MCEMfit_glm}}
}
\author{
Jakub Stoklosa and David I. Warton.
}
