% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.r
\docType{data}
\name{Priniadata}
\alias{Priniadata}
\title{The yellow-bellied Prinia \emph{Prinia flaviventris} capture-recapture data}
\format{
A data set that contains: 3 columns with 164 observations. The columns are defined as follows:
\describe{
\item{\code{w1}}{Bird wing lengths.}
\item{\code{cap}}{Number of times the individual was captured.}
\item{\code{noncap}}{Number of times the individual was not captured.}
}
}
\source{
See Hwang, Huang and Wang (2007) for full details on the data and study.
}
\usage{
Priniadata
}
\description{
Data set consisting of capture-recapture histories 164 uniquely captured birds across 17 weekly capture occasions. Bird wing lengths were also measured in the study.
}
\examples{
# Load the data.

data(Priniadata)
}
\references{
Hwang, W. H., Huang, S. Y. H., and Wang, C. (2007). Effects of measurement error and conditional score estimation in capture--recapture models. \emph{Statistica Sinica}, \strong{17}, 301-316.
}
\keyword{datasets}
