% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.r
\name{sqrt.na}
\alias{sqrt.na}
\title{Function that replaces NA with zero for a matrix}
\usage{
\method{sqrt}{na}(x)
}
\arguments{
\item{x}{: a matrix}
}
\value{
\code{sqrt.na} returns a matrix.
}
\description{
This function replaces NA with zero for a matrix.
}
\author{
Jakub Stoklosa
}
