% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\name{MCEMfit_glm}
\alias{MCEMfit_glm}
\title{MCEMfit_glm}
\usage{
MCEMfit_glm(mod, family, sigma.sq.u, sigma.sq.e, B, epsilon = 1e-05,
  theta.est = 1)
}
\arguments{
\item{mod}{:}

\item{family}{:}

\item{sigma.sq.u}{:}

\item{sigma.sq.e}{:}

\item{B}{:}

\item{epsilon}{:}

\item{theta.est}{:}
}
\value{
\code{MCEMfit_glm} returns a list of model coef estimates with standard errors.
}
\description{
Function for fitting the MCEM algorithm for non-Gaussian (GLMs) data with measurement error.
}
\author{
Jakub Stoklosa and David I. Warton
}
