% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\name{refitME}
\alias{refitME}
\title{refitME}
\usage{
refitME(mod, sigma.sq.u, B, epsilon = 1e-05)
}
\arguments{
\item{mod}{:}

\item{sigma.sq.u}{:}

\item{B}{:}

\item{epsilon}{:}
}
\value{
\code{refitME} returns returns a list of model coef estimates with standard errors.
}
\description{
Function that extracts the model object and wraps the MCEM algorithm (currently available for lm, glm and gam).
}
\author{
Jakub Stoklosa and David I. Warton
}
