% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\name{refitME}
\alias{refitME}
\title{refitME}
\source{
See \url{https://github.com/JakubStats/refitME} for an RMarkdown tutorial with examples.
}
\usage{
refitME(mod, sigma.sq.u, W, B = 50, epsilon = 1e-05, se.comp = TRUE)
}
\arguments{
\item{mod}{: a glm/gam object (this is the naive fitted model). Make sure the first input predictor variables are the selected error-contaminated variable (i.e., the W's).}

\item{sigma.sq.u}{: measurement error variance. A scalar if there is only one error-contaminated variable, otherwise must be a covariance matrix}

\item{W}{: a matrix of error-contaminated covariates.}

\item{B}{: the number of Monte Carlo replication values (default is set 50).}

\item{epsilon}{: convergence threshold (default is set to 0.00001).}

\item{se.comp}{: a logical to indicate if standard errors for model parameters should be calculated and returned (default is set to TRUE).}
}
\value{
\code{refitME} returns model coef estimates with standard errors and the effective sample size.
}
\description{
Function that extracts the fitted (naive) model object and wraps the MCEM algorithm to correct for measurement error/error-in-variables (currently available for lm, glm and gam, excludes quassi models).
}
\references{
Stoklosa, J., Hwang, W-H., and Warton, D.I. A general algorithm for error-in-variables modelling using Monte Carlo expectation maximization.
}
\seealso{
\code{\link{MCEMfit_glm}} and \code{\link{MCEMfit_gam}}
}
\author{
Jakub Stoklosa and David I. Warton.
}
