% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitME_package.R
\name{wt.var}
\alias{wt.var}
\title{Function that calculates a weighted variance}
\source{
The developer of this function is Jeremy VanDerWal. See \url{https://rdrr.io/cran/SDMTools/src/R/wt.mean.R}
}
\usage{
wt.var(x, w)
}
\arguments{
\item{x}{: a vector of numerical data.}

\item{w}{: a vector of equal length to \code{x} representing the weights.}
}
\value{
\code{wt.var} returns a single value from analysis requested.
}
\description{
This function that calculates a weighted variance for a given vector.
}
\examples{
# Define simple data
x = 1:25 # Set of numbers.
wt = runif(25) # Some arbitrary weights.

# Display variances (unweighted and then weighted).
var(x)
wt.var(x, wt)
}
